Reimplement intrinsic event iteration.
The new code processes any attribute that starts with "on", or href attributes that start with "javascript:" and parses them assigning them a unique url, so that all of them are independently addressable by the panel and by white/blacklisting. This fixes the fact that the previous code would only parse some types of attributes using a list of names, would only parse the first matching attribute in an element, and would only list one intrinsic event per page in the panel. It also fixes the content of the attribute being mangled to the first character of the original code.