Fix infinite loop
LibreJS got stuck in an infinite loop whenever a

    <link rel="jslicense">

occured inside a HTML head tag.  This was due to a small programming
mistake (refer to the git diff for details).  It caused Icecat to become
unresponsive (tested on Icecat 78.10.0esr).

This notably affected users trying to visit lichess.org, and reportedly
users trying to visit scratch.mit.edu [0].

[0]: https://lists.gnu.org/archive/html/bug-librejs/2020-06/msg00000.html